rename diagram modal
