+++ packages/frinx-workflow-ui/src/pages/workflow-list/WorkflowDefs/diagram-modal/diagram-modal.tsx
