+++ packages/frinx-workflow-ui/src/pages/workflow-list/WorkflowDefs/workflow-defs-read-only.js
@@ -11,7 +11,7 @@
 // @flow
 import DefinitionModal from './definition-modal/definition-modal';
 import DependencyModal from './DependencyModal/DependencyModal';
-import DiagramModal from './DiagramModal/DiagramModal';
+import DiagramModal from './diagram-modal/diagram-modal';
 import InputModal from './InputModal/input-modal';
 import PageContainer from '../../../common/PageContainer';
 import Paginator from '../../../common/pagination';
